runs codegen on build & remove stopWatcher console log
