+++ .changeset/green-pillows-hammer.md
@@ -1,0 +1,5 @@
+---
+"blitz": patch
+---
+
+Runs the codegen on the blitz build command
