+++ packages/blitz/src/cli/utils/next-commands.ts
@@ -9,10 +9,11 @@ import {
 } from "./next-utils"
 import {checkLatestVersion} from "./check-latest-version"
 import {readBlitzConfig} from "../../server-utils"
+import {codegenTasks} from "./codegen-tasks"
 
 export async function build(config: ServerConfig) {
   const {rootFolder, nextBin, watch} = await normalize(config)
-
+  await codegenTasks()
   await nextBuild(nextBin, rootFolder, {} as any, config)
   if (customServerExists()) await buildCustomServer({watch})
 }
